% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrenceNetwork.R
\name{cooccurrenceNetwork}
\alias{cooccurrenceNetwork}
\title{Plots a co-occurrence network visualizing the co-occurrence of taxa (at genus level) in spots. Edges are drawn if SparCC correlation > threshold.}
\usage{
cooccurrenceNetwork(
  object,
  threshold = 0.1,
  taxonomizrDB = "/projects/site/pred/microbiome/database/taxonomizr_DB/nameNode.sqlite"
)
}
\arguments{
\item{object}{List. Output from decontaminate() function.}

\item{threshold}{double. SparCC correlation threshold (value between -1 and 1) to be used to define edges between taxa on co-occurrence network. Defaults to 0.1.}

\item{taxonomizrDB}{A string. Path to nameNode.sqlite database required for taxonomic conversions (see README file for how to download).}
}
\value{
two plots: first plot is an igraph plot, second plot is the corresponding legend.
}
\description{
Plots a co-occurrence network visualizing the co-occurrence of taxa (at genus level) in spots. Edges are drawn if SparCC correlation > threshold.
}
\examples{
\dontrun{
cooccurrenceNetwork(CRC_16, threshold=0.1)
cooccurrenceNetwork(CRC_16, threshold=0.5)
}

}
