% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CRC_16}
\alias{CRC_16}
\title{CRC_16 toy data}
\format{
\subsection{\code{CRC_16}}{

A list with 3 objects: $seurat_object, $taxid_counts, $matrix
\describe{
\item{CRC_16$seurat_object}{An object of class seurat with two assays: GEX (host transcriptomics) and TAXA_G (microbiome)}
\item{CRC_16$taxid_counts}{A dataframe with the column counts, superkingdom, phylum, genus and taxid}
\item{CRC_16$matrix}{A sparce Matrix of class "dgCMatrix", with rows=taxa and columns=spots}
}
}
}
\source{
\url{https://doi.org/10.1038/s41586-022-05435-0}
}
\usage{
CRC_16
}
\description{
Example output of the decontaminate() function to be used for testing of
downstream analyis and visualization, such as spatialPlot(),
pseudoBulkProfile() or cooccurrenceNetwork().
}
\details{
Original data from Galeano-Niño et al. (Nature 2022) publication.
Processed with Roche bioinformatic pipeline and default settings in
microbiome10XVisium krakenToMatrix() and decontaminate().
}
\keyword{datasets}
