% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontaminate.R
\name{decontaminate}
\alias{decontaminate}
\title{Performs various decontamination steps on the taxid-spot matrix produced by krakenToMatrix().}
\usage{
decontaminate(
  sampleName,
  filePath = NULL,
  object = NULL,
  spacerangerDir,
  outDir,
  removeSingletons = TRUE,
  removeLikelyContaminants = TRUE,
  selectGastrointestinal = FALSE,
  spots = "tissueOnly",
  distance = 7,
  removeSpecificTaxa = NULL,
  taxonomizrDB = "/projects/site/pred/microbiome/database/taxonomizr_DB/nameNode.sqlite"
)
}
\arguments{
\item{sampleName}{A string. Sample name.}

\item{filePath}{A string. Path to RDS object created with krakenToMatrix(). (note: either filePath or object has to be supplied)}

\item{object}{A list. Name of the R object from the output of krakenToMatrix(). (note: either filePath or object has to be supplied)}

\item{spacerangerDir}{A string. Path to the spaceranger outs folder.}

\item{outDir}{A string. Path to desired output directory. RDS object will be saved as sampleName_Seurat_object.RDS.}

\item{removeSingletons}{A logical. Whether to remove counts==1 in spots. Defaults to TRUE.}

\item{removeLikelyContaminants}{A logical. Whether to remove taxa defined as likely contaminants by Poore et al. (2020 Nature). Only applicable if tax_level=="genus" in krakenToMatrix. Defaults to TRUE.}

\item{selectGastrointestinal}{A logical. Whether to only keep taxa defined as oral, fecal or oral/fecal transmitter by Schmidt et al. (2019 eLife). Only applicable if tax_level=="genus" in krakenToMatrix. Defaults to FALSE.}

\item{spots}{A string. Which spots to consider, either one of c("tissueOnly", "tissuePlusBordering", "all"). Defaults to "tissueOnly".}

\item{distance}{An integer. Only required when spots=="tissuePlusBordering". Determines the size of the bordering region surrounding the tissue. Recommended values in range of (2,7). Defaults to 7.}

\item{removeSpecificTaxa}{A vector of strings. Vector of genus names of taxa to remove. Defaults to NULL.}

\item{taxonomizrDB}{A string. Path to nameNode.sqlite database required for taxonomic conversions (see README file for how to download).}
}
\value{
A list of $seurat_object(Seurat object with decontaminated taxa assay),$taxid_counts (dataframe with taxids, taxonomic names and counts) and $matrix (taxid-spot matrix).
}
\description{
Performs various decontamination steps on the taxid-spot matrix produced by krakenToMatrix().
}
\examples{
\dontrun{
 decontaminate(sampleName = "CRC_16",
 filePath=system.file("extdata", "CRC_16", "genus_umi_counts.RDS", package="microbiome10XVisium"),
 spacerangerDir = system.file("extdata", "outs/", package="microbiome10XVisium"),
 outDir = system.file("extdata", "CRC_16/", package="microbiome10XVisium"))
}
}
