% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontaminateSingleCell.R
\name{decontaminateSingleCell}
\alias{decontaminateSingleCell}
\title{Performs various decontamination steps for scRNA-seq datasets on the taxid-spot matrix produced by krakenToMatrix().}
\usage{
decontaminateSingleCell(
  sampleName,
  filePath = NULL,
  object = NULL,
  outDir,
  removeSingletons = TRUE,
  removeLikelyContaminants = TRUE,
  selectGastrointestinal = TRUE,
  removeSpecificTaxa = NULL,
  taxonomizrDB = "/projects/site/pred/microbiome/database/taxonomizr_DB/nameNode.sqlite"
)
}
\arguments{
\item{sampleName}{A string. Sample name.}

\item{filePath}{A string. Path to RDS object created with krakenToMatrix(). (note: either filePath or object has to be supplied)}

\item{object}{A list. Name of the R object from the output of krakenToMatrix(). (note: either filePath or object has to be supplied)}

\item{outDir}{A string. Path to output directory. RDS object saved as sampleName_decontaminated.RDS.}

\item{removeSingletons}{A logical. Whether to remove umi_counts==1 in spots and prevalence(taxid)==1. Only applicable if counts=="umi_counts" in krakenToMatrix(). Defaults to TRUE.}

\item{removeLikelyContaminants}{A logical. Whether to remove taxa defined as likely contaminants by Poore et al. (Nature 2020). Only applicable if tax_level=="genus" in krakenToMatrix(). Defaults to TRUE.}

\item{selectGastrointestinal}{A logical. Whether to only keep taxa defined as oral, fecal or oral/fecal transmitter by Schmidt et al. (2019 eLife). Only applicable if tax_level=="genus" in krakenToMatrix. Defaults to FALSE.}

\item{removeSpecificTaxa}{A vector of strings. Vector of genus names of taxa to remove. Defaults to NULL.}

\item{taxonomizrDB}{A string. Path to nameNode.sqlite database required for taxonomic conversions (see README file for how to download).}
}
\value{
A list of $matrix (decontaminated taxid-cell matrix) and $taxid_counts (dataframe with taxids, taxa name and counts).
}
\description{
Performs various decontamination steps for scRNA-seq datasets on the taxid-spot matrix produced by krakenToMatrix().
}
\examples{
\dontrun{
decontaminateSingleCell(sampleName="P1_D0", filePath="FILEPATH",
outDir="OUTDIR", removeSpecificTaxa=c("Mycobacterium"))
}

}
