% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_matrix.R
\name{export_matrix}
\alias{export_matrix}
\title{Exports the decontaminated taxa-spot matrix to desired file format for external downstream use.}
\usage{
export_matrix(
  object,
  rownames = "genus",
  fileFormat = "csv",
  outDir,
  taxonomizrDB = "/projects/site/pred/microbiome/database/taxonomizr_DB/nameNode.sqlite"
)
}
\arguments{
\item{object}{List. Output from decontaminate() function.}

\item{rownames}{String. Whether rownames of matrix should be taxid ("taxid") or the taxa names at genus level ("genus"). Defaults to "genus".}

\item{fileFormat}{String. File format of exported matrix. Defaults to "csv".}

\item{outDir}{String. Path to desired output directory. Matrix will be saved as either "microbiome_genus_matrix" or "microbiome_taxid_matrix"}

\item{taxonomizrDB}{A string. Path to nameNode.sqlite database required for taxonomic conversions (see README file for how to download).}
}
\value{
none
}
\description{
Exports the decontaminated taxa-spot matrix to desired file format for external downstream use.
}
\examples{
export_matrix(object=CRC_16, rownames="genus", fileFormat="csv",
outDir=system.file("extdata", "CRC_16/", package="microbiome10XVisium"))

}
