% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krakenToMatrix.R
\name{krakenToMatrix}
\alias{krakenToMatrix}
\title{Creates a taxid-spot matrix from snakemake workflow output file.}
\usage{
krakenToMatrix(
  filePath,
  counts = "umi_counts",
  tax_level = "genus",
  whitelist = "10XVisium",
  taxonomizrDB = "/projects/site/pred/microbiome/database/taxonomizr_DB/nameNode.sqlite",
  outDir
)
}
\arguments{
\item{filePath}{A string. Path to input file (kraken output file from snakemake workflow: SAMPLE_profiling-output.txt).}

\item{counts}{A character string. Either "umi_counts" or "read_counts". Defaults to "umi_counts".}

\item{tax_level}{A character string. Either "species", "genus" or "family". It is recommended to use "genus". Defaults to "genus".}

\item{whitelist}{A character string. Either "10XVisium" (when working with spaceranger output data) or "10XChromium_v3.1" (when working with cellranger and Single Cell 3' v3.1 technology data).}

\item{taxonomizrDB}{A string. Path to nameNode.sqlite database required for taxonomic conversions (see README file for how to download).}

\item{outDir}{A character string. Path to desired output directory.}
}
\value{
A list containing $taxid_counts (dataframe with counts, taxonomic names and taxids) and $matrix (matrix with rows=taxids, cols=barcodes).
}
\description{
Creates a taxid-spot matrix from snakemake workflow output file.
}
\examples{
#for spatial transcriptomics
\dontrun{
 krakenToMatrix(filePath=
 system.file("extdata", "CRC_16", "CRC_16_profiling-output.txt.gz", package="microbiome10XVisium"),
 outDir=system.file("extdata", "CRC_16/", package="microbiome10XVisium"))
}
#for scRNAseq
\dontrun{
 krakenToMatrix(filePath="FILEPATH", outDir="OUTDIR", whitelist="10XChromium_v3.1")
}
}
