% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoBulkProfile.R
\name{pseudoBulkProfile}
\alias{pseudoBulkProfile}
\title{Creates a microbiome relative abundance bar plot of the decontaminated sample (at genus level, showing the top 10 taxa).}
\usage{
pseudoBulkProfile(sampleName, object)
}
\arguments{
\item{sampleName}{String. Sample name.}

\item{object}{List. Output from decontaminate() function.}
}
\value{
ggplot2 plot.
}
\description{
Creates a microbiome relative abundance bar plot of the decontaminated sample (at genus level, showing the top 10 taxa).
}
\examples{
pseudoBulkProfile(sampleName="CRC_16", object=CRC_16)
}
