% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{spatialPlot}
\alias{spatialPlot}
\title{Creates a Seurat SpatialFeaturePlot visualizing the number of UMIs or reads of taxa on the tissue image.}
\usage{
spatialPlot(
  object,
  taxa,
  taxonomizrDB = "/projects/site/pred/microbiome/database/taxonomizr_DB/nameNode.sqlite"
)
}
\arguments{
\item{object}{A list. Output from decontaminate() function.}

\item{taxa}{A string. Either one of "all", "bacterial", "eukaryotic", "viral", "nTaxa" (i.e. indicating the number of different taxa detected per spot) or a taxa name at genus level, e.g. "Fusobacterium".}

\item{taxonomizrDB}{A string. Path to nameNode.sqlite database required for taxonomic conversions (see README file for how to download).}
}
\value{
a Seurat::SpatialFeaturePlot
}
\description{
Creates a Seurat SpatialFeaturePlot visualizing the number of UMIs or reads of taxa on the tissue image.
}
\examples{
\dontrun{
spatialPlot(CRC_16, taxa="Fusobacterium")
spatialPlot(CRC_16, taxa="all")
spatialPlot(CRC_16, taxa="eukaryotic")
spatialPlot(CRC_16, taxa="nTaxa")
}
}
