% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_spatialPlot.R
\name{interactive_spatialPlot}
\alias{interactive_spatialPlot}
\title{Creates an interactive Seurat SpatialFeaturePlot allowing to probe for different taxa and visualize the UMI counts of taxa on the tissue image.}
\usage{
interactive_spatialPlot(object)
}
\arguments{
\item{object}{List. Output from decontaminate() function, containing $seurat_object and $taxid_counts.}
}
\value{
a Seurat::SpatialFeaturePlot
}
\description{
Creates an interactive Seurat SpatialFeaturePlot allowing to probe for different taxa and visualize the UMI counts of taxa on the tissue image.
}
\examples{
if (FALSE){
 interactive_spatialPlot(CRC_16)
}

}
